requirements.txt

django>=4.2.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.0
python-decouple>=3.8
psycopg2-binary>=2.9
